fix php 8.5 issues (#214)

+++ Helper/Data.php
@@ -494,7 +494,7 @@ class Data extends AbstractHelper
      */
     public function truncateDecimal($amount)
     {
-        return floor((double) $amount);
+        return floor((float) $amount);
     }
 
     /**
